Enfim API conectada

+++ app/src/main/java/com/example/android/siv/Aplicador.kt
@@ -7,5 +7,5 @@ data class Aplicador (
 	@SerializedName("aplicadorId") val aplicadorId : Int,
 	@SerializedName("crmCreAplicador") val crmCreAplicador : Int,
 	@SerializedName("nomeAplicador") val nomeAplicador : String,
-	@SerializedName("telefoneAplicador") val telefoneAplicador : Int
+	@SerializedName("telefoneAplicador") val telefoneAplicador : Long
 )
